병충해 백/프론트 응답형식 동기화, 예외처리수정, 프론트 더미데이터 제거
