--- backend/matefarm/src/main/java/com/avengers/matefarm/insectpestinfo/common/ApiCode.java
@@ -1,5 +1,0 @@
-package com.avengers.matefarm.insectpestinfo.common;
-
-public enum ApiCode {
-    SU, NF, ER
-}
