+++ backend/matefarm/src/main/java/com/avengers/matefarm/insectpestinfo/service/DiseaseDetailService.java
@@ -1,9 +1,14 @@
-package com.avengers.matefarm.insectpestinfo.disease.service;
+package com.avengers.matefarm.insectpestinfo.service;
 
-import com.avengers.matefarm.insectpestinfo.disease.dto.*;
-import com.avengers.matefarm.insectpestinfo.disease.repository.DiseaseDetailRepository;
-import com.avengers.matefarm.insectpestinfo.disease.repository.DiseaseRepository;
+import com.avengers.matefarm.common.exception.CommonException;
+import com.avengers.matefarm.common.exception.ErrorCode;
+import com.avengers.matefarm.insectpestinfo.dto.DiseaseBaseDto;
+import com.avengers.matefarm.insectpestinfo.dto.DiseaseDetailDto;
+import com.avengers.matefarm.insectpestinfo.dto.DiseaseDetailResponse;
 import com.avengers.matefarm.insectpestinfo.entity.DiseaseEntity;
+import com.avengers.matefarm.insectpestinfo.repository.DiseaseDetailRepository;
+import com.avengers.matefarm.insectpestinfo.repository.DiseaseRepository;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,7 +23,7 @@ public class DiseaseDetailService {
     @Transactional(readOnly = true)
     public DiseaseDetailResponse get(Long diseaseId) {
         DiseaseEntity base = diseaseRepository.findByIdAndStatus(diseaseId, 1)
-                .orElseThrow(() -> new IllegalArgumentException("NF"));
+                .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_POST));
 
         var detail = diseaseDetailRepository.findById(diseaseId).orElse(null);
